fix: 온보딩 지역 및 학교 전공 검색 보완

+++ src/main/java/com/wishconnect/domain/user/dto/request/ProfileHouseholdRequest.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record ProfileHouseholdRequest(
 		@NotBlank String incomeLevel,
-		@Min(1) Integer familySize,
+		@Min(1) Long familySize,
 		List<String> familyTypes,
 		List<String> personalStatuses,
 		List<String> interests
